Expanded chimera test coverage
